Add support for .dockerignore, .gcloudignore and .gitignore
New ignore files have been added to reduce the size of Docker and Google Cloud builds, and to specify intentionally untracked files in Git that should be ignored. A new Cargo.lock file has also been generated to ensure that everyone working on the project will use the same versions of all dependencies.