Update package versions, improve error handling and image processing
The package versions in Cargo were updated, notably the 'rocket' and 'tokio' versions. Error handling for unzip function in 'download.rs' was enhanced. Image processing was improved via changes in the 'contrast' logic in 'main.rs', and changes for channels needed for color composition. Token refreshing was added in 'authenticate.rs'. Dockerfile was also modified to include new steps for OpenCV installation.